+++ example/nodejs/src/context.ts
@@ -1,6 +1,6 @@
 import pino from 'pino';
 import { Sequelize } from 'sequelize';
-import { container } from '@viron/lib';
+import { repositoryContainer } from '@viron/lib';
 import { Mode, MODE_MONGO, MODE_MYSQL } from './constants';
 import { preflight as preflightMongo } from './stores/mongo';
 import { preflight as preflightMysql } from './stores/mysql';
@@ -87,7 +87,7 @@ export class Context {
         throw noSetEnvMode();
     }
 
-    container.init(this.mode, this.stores.main.instance);
+    repositoryContainer.init(this.mode, this.stores.main.instance);
   }
 }
