+++ example/nodejs/src/openapi/root.yaml
@@ -2,7 +2,10 @@ openapi: 3.0.2
 info:
   version: 0.0.1
   title: '@viron/example-nodejs'
-  description: '@viron/example-nodejs api specifications'
+  description: api specifications
+
+servers:
+  - url: /
 
 tags:
   - name: root
