+++ example/nodejs/src/routes/ping.ts
@@ -1,4 +1,4 @@
-import { container } from '@viron/lib';
+import { repositoryContainer } from '@viron/lib';
 import { getUserRepository } from '../repositories';
 import { RouteContext } from '.';
 
@@ -17,7 +17,7 @@ export const getPing = async (context: RouteContext): Promise<void> => {
   const fdoc1 = await getUserRepository().findOneById(doc1.id);
   console.log(fdoc1);
 
-  const doc2 = await container.getAuditLogRepository().createOne({
+  const doc2 = await repositoryContainer.getAuditLogRepository().createOne({
     requestMethod: 'GET',
     requestUri: `/ping?now=${now}`,
     sourceIp: '127.0.0.1',
@@ -26,7 +26,9 @@ export const getPing = async (context: RouteContext): Promise<void> => {
     statusCode: 200,
   });
   console.log('create', doc2);
-  const fdoc2 = await container.getAuditLogRepository().findOneById(doc2.id);
+  const fdoc2 = await repositoryContainer
+    .getAuditLogRepository()
+    .findOneById(doc2.id);
   console.log(fdoc2);
 
   context.res.setBody('pong');
