+++ packages/nodejs/src/domains/auditlog.ts
@@ -1,4 +1,4 @@
-import { container } from '../';
+import { repositoryContainer } from '../';
 
 export interface AuditLog {
   id: string;
@@ -22,6 +22,6 @@ export interface AuditLogCreationAttributes {
 }
 
 export const list = async (): Promise<AuditLog[]> => {
-  const repository = container.getAuditLogRepository();
+  const repository = repositoryContainer.getAuditLogRepository();
   return repository.find();
 };
